pyproject.toml
[project]
name = "costcut"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.1.0",
    "boto3>=1.42.30",
    "croniter>=6.0.0",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.1.0",
    "django-environ>=0.12.0",
    "django-q2>=1.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.8",
    "drf-spectacular>=0.29.0",
    "google-cloud-billing>=1.18.0",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "gunicorn>=23.0.0",
    "isort>=7.0.0",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg>=3.3.2",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.git
  | \.venv
  | venv
  | build
  | dist
  | migrations
  | staticfiles
  | media
)/
'''


[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["*/migrations/*", "*/staticfiles/*", "*/media/*"]
known_first_party = ["apps", "costcut"]
known_django = ["django"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "costcut.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
testpaths = ["apps"]
